Move the base provider classes into a Base module
This way it is clear which code needs to be inherited from when creating
a new provider
